Move __main__ contents to separate module
